+++ demisto_sdk/tests/integration_tests/validate_integration_test.py
@@ -3494,7 +3494,7 @@ class TestSpecificValidations:
         - Running validate on it with flag --run-specific-validations RP101.
 
         Then
-        - Ensure validate fails on RP101 - wrong value in expiration field and not on RP102 - id and details fields are not equal..
+        - Ensure validate fails on RP101 - wrong value in expiration field and not on RP102 - id and details fields are not equal.
         """
         mocker.patch.object(tools, 'is_external_repository', return_value=True)
         pack = repo.create_pack('PackName')
@@ -3508,4 +3508,31 @@ class TestSpecificValidations:
         assert f'Validating {reputation.path} as reputation' in result.stdout
         assert 'RP101' in result.stdout
         assert 'Expiration field should have a positive numeric value.' in result.stdout
+        assert result.exit_code == 1
+
+    def test_validate_with_flag_specific_validation_entire_code_section(self, mocker, repo):
+        """
+        Given
+        - an invalid Reputation - negative integer in expiration field and a 'details' that does not match its id.
+
+        When
+        - Running validate on it with flag --run-specific-validations RP.
+
+        Then
+        - Ensure validate fails on RP101 - wrong value in expiration field and on RP102 - id and details fields are not equal.
+        """
+        mocker.patch.object(tools, 'is_external_repository', return_value=True)
+        pack = repo.create_pack('PackName')
+        reputation_copy = REPUTATION.copy()
+        reputation_copy['expiration'] = -1
+        reputation_copy["details"] = "reputationn"
+        reputation = pack._create_json_based(name='reputation', prefix='', content=reputation_copy)
+        with ChangeCWD(pack.repo_path):
+            runner = CliRunner(mix_stderr=False)
+            result = runner.invoke(main, [VALIDATE_CMD, '-i', reputation.path, '--run-specific-validations', 'RP'], catch_exceptions=False)
+        assert f'Validating {reputation.path} as reputation' in result.stdout
+        assert 'RP101' in result.stdout
+        assert 'Expiration field should have a positive numeric value.' in result.stdout
+        assert 'RP102' in result.stdout
+        assert 'id and details fields are not equal.' in result.stdout
         assert result.exit_code == 1